Good update
This Update:
* You can now download the pictures in your notes (took forever, but it converts the blobs to objects, and then passes the data uri which android can handle)
* Floating circle New Note button is back (Microsoft fixed the overlap bug)
* Fixed image rotation bug when uploading from camera (exif data wasn't being read properly with Sticky Notes)
* Updated dark theme a bit (image menu/hover fix)
* Removed an app permission (don't need the extend permission after all)

Last Update:
* Fixed issue w/ camera save location on Android 10

Note: Microsoft has been putting in a few changes here and there recently, such as part of the theme, and the now new bug with picture 'right shift' bug when viewing a note (but I bet it won't last long anyways). Update to the Gboard and SwiftKey users, I have seen some promising changes with how Sticky Notes accepts text, but for now, it's still not working 100% (text is no longer jumpy/breaks, but it doesn't sync text).